requirements.txt
# ------------------------------ #

sanic
sanic-ext

# ------------------------------ #

pyjwt

# ------------------------------ #

peewee

# ------------------------------ #

marshmallow

# ------------------------------ #
